requirements.txt
fastapi==0.58.1
pydantic==1.5.1
SQLAlchemy==1.3.18
alembic==1.4.2
psycopg2==2.7.5

python-dotenv==0.14.0
uvicorn==0.11.5
gunicorn==19.9.0

line-bot-sdk==1.8.0
requests==2.20.0
beautifulsoup4==4.9.1
pytz==2018.6
